Some speedups on string and boolean encoding primarily by reducing the amount of allocations required, thus reducing gc garbage as well.


String results:
```
$ benchcmp a.txt b.txt 
benchmark                              old ns/op     new ns/op     delta
BenchmarkStringSerialization_small     157           123           -21.66%
BenchmarkStringSerialization_med       181           126           -30.39%
BenchmarkStringSerialization_large     730           186           -74.52%
```

Boolean results:
```
$ benchcmp a.txt b.txt 
benchmark                           old ns/op     new ns/op     delta
BenchmarkBooleanSerialization-4     36.3          21.8          -39.94%
```